requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
python-jose
passlib[bcrypt]
pymysql
python-dotenv
firebase-admin